Create setupSearchHotkeys.js
